+++ stdlib/public/runtime/Private.h
@@ -172,18 +172,6 @@ namespace swift {
       dest = newValue;
   }
 
-#if defined(__CYGWIN__) || defined(_MSC_VER)
-  struct dl_phdr_info {
-    void *dlpi_addr;
-    const char *dlpi_name;
-  };
-
-  int _swift_dl_iterate_phdr(int (*Callback)(struct dl_phdr_info *info,
-                                             size_t size, void *data),
-                             void *data);
-  uint8_t *_swift_getSectionDataPE(void *handle, const char *sectionName,
-                                   unsigned long *sectionSize);
-#endif
 #if defined(__CYGWIN__)
   void _swift_once_f(uintptr_t *predicate, void *context,
                      void (*function)(void *));
